Created a separate component for the InitiativeOrderAccordion for organization
